Force web element lookups on attribute access, even if the element has already been located.
Not the most efficient, but the assumption should be that if a named Locator is accessed via attribute, it should be looked up EVERY time.  This means that if you only want to have selentric look something up the first time it's accessed via attribute, you want to store the result and work with the result via variable i.e. `divs = some_page.divs`